+++ scripts/update-package-versions.sh
@@ -3,19 +3,29 @@
 
 set -euo pipefail   # Bash "strict mode"
 script_dirpath="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
-
+root_dirpath="$(dirname "${script_dirpath}")"
 
 
 # ==================================================================================================
 #                                             Constants
 # ==================================================================================================
-LANG_ROOT_DIRPATH="$(dirname "${script_dirpath}")"
+SUPPORTED_LANGS_FILENAME="supported-languages.txt"
+UPDATE_VERSION_IN_FILE_SCRIPT_FILENAME="update-version-in-file.sh"  # From devtools; expected to be on PATH
+SKIP_KEYWORD="SKIP" # Certain languages don't need any package replacement; they will use this keyword for their package file and replacement pattern
 
-SED_REPLACE_FILENAME_SUFFIX=".DELETEME"
+# Per-language package filepaths that need version, RELATIVE TO THE LANG ROOT, that need version replacement
+declare -A PACKAGE_FILE_RELATIVE_FILEPATHS
 
-SPECIFIED_FILENAME="package.json"
 
+declare -A PACKAGE_VERSION_PRINTF_PATTERNS
 
+# Golang -- empty because Go doesn't need any package version replacing
+PACKAGE_FILE_RELATIVE_FILEPATHS["golang"]="${SKIP_KEYWORD}"
+PACKAGE_VERSION_PRINTF_PATTERNS["golang"]="${SKIP_KEYWORD}"
+
+# Typescript
+PACKAGE_FILE_RELATIVE_FILEPATHS["typescript"]="package.json"
+PACKAGE_VERSION_PRINTF_PATTERNS["typescript"]="\"version\": \"%s\""
 
 # ==================================================================================================
 #                                       Arg Parsing & Validation
@@ -31,7 +41,7 @@ show_helptext_and_exit() {
 new_version="${1:-}"
 
 if [ -z "${new_version}" ]; then
-    echo "Error: no new version provided" >&2
+    echo "Error: No new version provided" >&2
     show_helptext_and_exit
 fi
 
@@ -40,32 +50,33 @@ fi
 # ==================================================================================================
 #                                             Main Logic
 # ==================================================================================================
-function getSearchReplace() {
-     value="${1:-}"
-     echo "\"version\": \"${value}\""
-}
 
+echo "Updating versions in package files for all supported langauges..."
+supported_langs_filepath="${root_dirpath}/${SUPPORTED_LANGS_FILENAME}"
+for lang in $(cat "${supported_langs_filepath}"); do
+    to_update_relative_filepath="${PACKAGE_FILE_RELATIVE_FILEPATHS["${lang}"]}"
+    if [ -z "${to_update_relative_filepath}" ]; then
+        echo "Error: No relative filepath to a package file that needs updating was found for language '${lang}'; this script needs to be updated with this information" >&2
+        exit 1
+    fi
+    if [ "${to_update_relative_filepath}" == "${SKIP_KEYWORD}" ]; then
+        continue
+    fi
 
-# Check that our package.json can be found in the right location
-if ! find ${LANG_ROOT_DIRPATH} -name ${SPECIFIED_FILENAME}; then
-    echo "Error: No ${SPECIFIED_FILENAME} file can be found inside '${LANG_ROOT_DIRPATH}'" >&2
-    exit 1
-fi
+    to_update_abs_filepath="${root_dirpath}/${lang}/${to_update_relative_filepath}"
 
-#Check that the package.json contains exactly one line matching "version": "X.Y.Z"
-if test $(grep -c $(getSearchReplace [0-9]\+\.[0-9]\+\.[0-9]\+) "${LANG_ROOT_DIRPATH}/${SPECIFIED_FILENAME}") -eq 1; then
-    echo "Error: Couldn't find one line matching \"version\": \"X.Y.Z\" in ${SPECIFIED_FILENAME}" >&2
-    exit 1
-fi
+    replacement_pattern="${PACKAGE_VERSION_PRINTF_PATTERNS["${lang}"]}"
+    if [ -z "${replacement_pattern}" ]; then
+        echo "Error: No replacement pattern was found for language '${lang}'; this script needs to be updated with this information" >&2
+        exit 1
+    fi
+    if [ "${replacement_pattern}" == "${SKIP_KEYWORD}" ]; then
+        continue
+    fi
 
-# Do the replacement on the file - in this case, on package.json
-if ! sed -i "${SED_REPLACE_FILENAME_SUFFIX}" -E "s/$(getSearchReplace [0-9]\+\.[0-9]\+\.[0-9]\+)/$(getSearchReplace ${new_version})/g" ${SPECIFIED_FILENAME}; then #TODO No magic values, but can I use regex in variables?
-    echo "Error: Replacing pattern '${search_pattern}' with value '${new_version}' in file ${SPECIFIED_FILENAME}" >&2
-    exit 1
-fi
-
-# Delete the backup files that sed created
-if ! find "${LANG_ROOT_DIRPATH}" -name "*${SED_REPLACE_FILENAME_SUFFIX}" -delete; then
-    echo "Error: Couldn't delete the backup files that sed created ending in suffix '${SED_REPLACE_FILENAME_SUFFIX}'" > &2
-    exit 1
-fi
+    if ! "${UPDATE_VERSION_IN_FILE_SCRIPT_FILENAME}" "${to_update_abs_filepath}" "${replacement_pattern}" "${new_version}"; then
+        echo "Error: An error occurred setting new version '${new_version}' in '${lang}' package file '${to_update_abs_filepath}' using pattern '${replacement_pattern}'" >&2
+        exit 1
+    fi
+done
+echo "Successfully updated the versions in the package files for all supported languages"