+++ docs/changelog.md
@@ -1,6 +1,7 @@
 # TBD
 ### Features
 * Added index.ts to successfully export all necessary typescript code to npm
+* Added an `update-package-versions.sh` script that's hooked into the release flow for automatically updating versions in package files (e.g. `package.json`)
 
 ### Changes
 * Exported ContainerCreationConfigBuilder and ContainerRunConfigBuilder so that users can access them in the npm package